xxhsum: move programs/xxhsum -> cli by request
also use XXHSUM_SRC_DIR in Makefile
